# openapi.yaml
openapi: 3.1.0
info:
  title: Dash API
  version: 0.1.0
  description: >
    Single source of truth for the Dash API. The TypeScript client in
    packages/api-client is generated from this file — do not hand-write
    API types in the frontend.
servers:
  - url: /
paths:
...
  /api/systems:
    get:
      operationId: listSystems
      responses:
        "200":
          description: All monitored systems with their latest sample
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/System"
    post:
      operationId: createSystem
      summary: Register a system; the returned token authenticates agent pushes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "201":
          description: Created system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/systems/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      operationId: getSystem
      responses:
        "200":
          description: System detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchSystem
      summary: Rename or rotate the agent token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                rotateToken: { type: boolean }
      responses:
        "200":
          description: Updated system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/System"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteSystem
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/systems/{id}/stats:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      operationId: systemStats
      summary: Stored agent samples for charting
      parameters:
        - name: hours
          in: query
          schema: { type: integer, default: 24, maximum: 168 }
      responses:
        "200":
          description: Samples oldest-to-newest
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/StatSample"
                    - type: object
                      properties:
                        ts: { type: string, format: date-time }
        "404":
          $ref: "#/components/responses/NotFound"
  /api/systems/ingest:
    post:
      operationId: ingestSystem
      summary: "Agent push endpoint; Authorization: Bearer <system token>"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatSample"
      responses:
        "200":
          description: Sample stored
        "401":
          description: Missing or unknown token
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/import:
    post:
      operationId: importBoard
      summary: Replace board state from an export payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Export"
      responses:
        "200":
          description: Imported board tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Section"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Url:
      type: object
      required: [id, url, label]
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        label:
          type: string
          description: Tag such as "local" / "external" or a custom label
        position:
          type: number
...
    Item:
      type: object
      required: [id, sectionId, kind, name, position, urls]
      properties:
        id:
          type: string
        sectionId:
          type: string
        kind:
          type: string
          enum: [service, widget]
        name:
          type: string
        icon:
          type: string
          description: "/api/icons/<file> path or remote URL; empty = letter tile"
        position:
          type: number
        config:
          type: object
          nullable: true
          additionalProperties: true
          description: Widget config (Phase 2); null for services
        urls:
          type: array
          items:
            $ref: "#/components/schemas/Url"
...
    Section:
      type: object
      required: [id, name, position, collapsed, boardId, items]
      properties:
        id:
          type: string
        name:
          type: string
        position:
          type: number
        collapsed:
          type: boolean
        boardId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    Board:
      type: object
      required: [id, slug, name, position]
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        position:
          type: number
...
    StatSample:
      type: object
      description: One agent push; missing fields mean the platform can't provide them
      properties:
        host: { type: string }
        os: { type: string }
        arch: { type: string }
        cpuModel: { type: string }
        cores: { type: integer }
        intervalS: { type: integer }
        uptimeS: { type: number }
        cpu:
          type: number
          description: CPU usage percent 0-100
        memTotal: { type: number, description: bytes }
        memUsed: { type: number, description: bytes }
        swapTotal: { type: number, description: bytes }
        swapUsed: { type: number, description: bytes }
        diskTotal: { type: number, description: root filesystem bytes }
        diskUsed: { type: number, description: bytes }
        netRx: { type: number, description: bytes per second }
        netTx: { type: number, description: bytes per second }
        load1: { type: number }
        load5: { type: number }
        load15: { type: number }
        temps:
          type: object
          additionalProperties: { type: number }
          description: Sensor label -> celsius
        containers:
          type: array
          items:
            $ref: "#/components/schemas/ContainerStat"
        smart:
          type: array
          items:
            $ref: "#/components/schemas/SmartDisk"
        zfs:
          type: array
          items:
            $ref: "#/components/schemas/ZFSPool"
        gpu:
          type: array
          items:
            $ref: "#/components/schemas/GPUStat"
    ContainerStat:
      type: object
      properties:
        name: { type: string }
        state: { type: string }
        cpu: { type: number, description: "percent of one core, may exceed 100" }
        memUsed: { type: number, description: bytes }
        memLimit: { type: number, description: bytes }
    SmartDisk:
      type: object
      properties:
        device: { type: string }
        model: { type: string }
        passed: { type: boolean, nullable: true }
        tempC: { type: number }
    ZFSPool:
      type: object
      properties:
        name: { type: string }
        health: { type: string }
        size: { type: number, description: bytes }
        free: { type: number, description: bytes }
    GPUStat:
      type: object
      properties:
        name: { type: string }
        tempC: { type: number }
        utilPct: { type: number }
        memUsed: { type: number, description: bytes }
        memTotal: { type: number, description: bytes }
    System:
      type: object
      required: [id, name, token, status]
      properties:
        id: { type: string }
        name: { type: string }
        token:
          type: string
          description: Agent bearer token
        host: { type: string }
        os: { type: string }
        arch: { type: string }
        cpuModel: { type: string }
        cores: { type: integer }
        intervalS: { type: integer }
        status:
          type: string
          enum: [pending, up, down]
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
        latest:
          $ref: "#/components/schemas/StatSample"
        createdAt:
          type: string
          format: date-time
...
    Export:
      type: object
      required: [version, sections]
      properties:
        version:
          type: integer
        settings:
          type: object
          additionalProperties: true
        sections:
          type: array
          items:
            $ref: "#/components/schemas/Section"
        boards:
          type: array
          items:
            $ref: "#/components/schemas/Board"